refactored into translation function factory
